requirements.txt
argcomplete==3.1.4
cryptography==3.4.8
netmiko==4.3.0
paramiko==2.12.0
pexpect==4.9.0
pynetbox==7.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
setuptools==68.1.2
wheel==0.42.0
